refactor functionality into many files
